+++ iqs-20241111/alibabacloud_iqs20241111/models.py
@@ -385,13 +385,190 @@ class GenericSearchResult(TeaModel):
         return self
 
 
-class GenericSearchRequest(TeaModel):
+class AiSearchRequest(TeaModel):
+    def __init__(
+        self,
+        industry: str = None,
+        page: int = None,
+        query: str = None,
+        session_id: str = None,
+        time_range: str = None,
+    ):
+        self.industry = industry
+        self.page = page
+        # This parameter is required.
+        self.query = query
+        self.session_id = session_id
+        self.time_range = time_range
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.industry is not None:
+            result['industry'] = self.industry
+        if self.page is not None:
+            result['page'] = self.page
+        if self.query is not None:
+            result['query'] = self.query
+        if self.session_id is not None:
+            result['sessionId'] = self.session_id
+        if self.time_range is not None:
+            result['timeRange'] = self.time_range
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('industry') is not None:
+            self.industry = m.get('industry')
+        if m.get('page') is not None:
+            self.page = m.get('page')
+        if m.get('query') is not None:
+            self.query = m.get('query')
+        if m.get('sessionId') is not None:
+            self.session_id = m.get('sessionId')
+        if m.get('timeRange') is not None:
+            self.time_range = m.get('timeRange')
+        return self
+
+
+class AiSearchResponseBodyHeader(TeaModel):
+    def __init__(
+        self,
+        event: str = None,
+        event_id: str = None,
+        response_time: int = None,
+    ):
+        self.event = event
+        self.event_id = event_id
+        self.response_time = response_time
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.event is not None:
+            result['event'] = self.event
+        if self.event_id is not None:
+            result['eventId'] = self.event_id
+        if self.response_time is not None:
+            result['responseTime'] = self.response_time
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('event') is not None:
+            self.event = m.get('event')
+        if m.get('eventId') is not None:
+            self.event_id = m.get('eventId')
+        if m.get('responseTime') is not None:
+            self.response_time = m.get('responseTime')
+        return self
+
+
+class AiSearchResponseBody(TeaModel):
+    def __init__(
+        self,
+        header: AiSearchResponseBodyHeader = None,
+        payload: str = None,
+        request_id: str = None,
+    ):
+        self.header = header
+        self.payload = payload
+        # Id of the request
+        self.request_id = request_id
+
+    def validate(self):
+        if self.header:
+            self.header.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.header is not None:
+            result['header'] = self.header.to_map()
+        if self.payload is not None:
+            result['payload'] = self.payload
+        if self.request_id is not None:
+            result['requestId'] = self.request_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('header') is not None:
+            temp_model = AiSearchResponseBodyHeader()
+            self.header = temp_model.from_map(m['header'])
+        if m.get('payload') is not None:
+            self.payload = m.get('payload')
+        if m.get('requestId') is not None:
+            self.request_id = m.get('requestId')
+        return self
+
+
+class AiSearchResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: AiSearchResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = AiSearchResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
+class GenericAdvancedSearchRequest(TeaModel):
     def __init__(
         self,
         query: str = None,
+        session_id: str = None,
         time_range: str = None,
     ):
+        # This parameter is required.
         self.query = query
+        self.session_id = session_id
         self.time_range = time_range
 
     def validate(self):
@@ -405,6 +582,8 @@ class GenericSearchRequest(TeaModel):
         result = dict()
         if self.query is not None:
             result['query'] = self.query
+        if self.session_id is not None:
+            result['sessionId'] = self.session_id
         if self.time_range is not None:
             result['timeRange'] = self.time_range
         return result
@@ -413,6 +592,101 @@ class GenericSearchRequest(TeaModel):
         m = m or dict()
         if m.get('query') is not None:
             self.query = m.get('query')
+        if m.get('sessionId') is not None:
+            self.session_id = m.get('sessionId')
+        if m.get('timeRange') is not None:
+            self.time_range = m.get('timeRange')
+        return self
+
+
+class GenericAdvancedSearchResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: GenericSearchResult = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = GenericSearchResult()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
+class GenericSearchRequest(TeaModel):
+    def __init__(
+        self,
+        industry: str = None,
+        page: int = None,
+        query: str = None,
+        session_id: str = None,
+        time_range: str = None,
+    ):
+        self.industry = industry
+        self.page = page
+        # This parameter is required.
+        self.query = query
+        self.session_id = session_id
+        self.time_range = time_range
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.industry is not None:
+            result['industry'] = self.industry
+        if self.page is not None:
+            result['page'] = self.page
+        if self.query is not None:
+            result['query'] = self.query
+        if self.session_id is not None:
+            result['sessionId'] = self.session_id
+        if self.time_range is not None:
+            result['timeRange'] = self.time_range
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('industry') is not None:
+            self.industry = m.get('industry')
+        if m.get('page') is not None:
+            self.page = m.get('page')
+        if m.get('query') is not None:
+            self.query = m.get('query')
+        if m.get('sessionId') is not None:
+            self.session_id = m.get('sessionId')
         if m.get('timeRange') is not None:
             self.time_range = m.get('timeRange')
         return self