+++ iqs-20241111/ChangeLog.md
@@ -1,3 +1,13 @@
+2024-11-15 Version: 1.1.0
+- Support API AiSearch.
+- Support API GenericAdvancedSearch.
+- Update API GenericSearch: add param industry.
+- Update API GenericSearch: add param page.
+- Update API GenericSearch: add param sessionId.
+- Update API GenericSearch: update param query.
+- Update API GenericSearch: update param timeRange.
+
+
 2024-11-05 Version: 1.0.1
 - Generated python 2024-11-11 for IQS.